Lots of changes:
1. for all calls of split(","), added trim() for the calls on the split
parts. that way, using "x,y" or "x, y" makes no difference.

2. added Water Class, to be able to distinguish between different water
types in the future (perhaps).

2. implemented water consumption:
currently very close to food consumption.
There is a difference though, between the water days in the store view
and the water days in the world view. need to find out why.
ConsumeWater() and ConsumeFood() now contain a check whether the
expedition has actually any members left. 
Is unnecessary in the game since this is checked somewhere, just not in
the JUnit test case game setup. Still something amiss then.

3. moved hunger and thirst resistance to expedition instead of
FoodConsumerDelegate.
Extended FoodConsumer Interface because of that. Not sure the whole
interface is actually needed. A Modeler Class for Consumption might all
thats necessary.

4. added water days to the Macro Level UI, in addition to gold.

5. added JUnit Test Case for Water and Food Consumption.

6. Added MessengerService Class, moved Date and date formatting methods
out of AbstractLevel. Todo: Move all Messenging To this class. Also
update all calls to point to that class instead.